+++ app/api/graphql/app.py
@@ -3,28 +3,13 @@ from fastapi import FastAPI
 
 from strawberry.asgi import GraphQL
 
+from ..graphql.schema.queries import query
 
-@strawberry.type
-class User:
-    name: str
-    age: int
-
-
-@strawberry.type
-class Query:
-    @strawberry.field
-    def user(self) -> User:
-        return User(name="Patrick", age=100)
-
-
-schema = strawberry.Schema(query=Query)
+schema = strawberry.Schema(query= query)
 
 
 graphql_app = GraphQL(schema)
 
-
 app = FastAPI()
 
-app.add_route("/graphql", graphql_app)
-
-app.add_websocket_route("/graphql", graphql_app)
+app.include_router(graphql_app, prefix="/graphql")
