+++ app/api/graphql/schema/__init__.py
@@ -1,0 +1,1 @@
+from .types import *
