+++ app/api/graphql/schema/queries/__init__.py
@@ -1,0 +1,1 @@
+from .query import *
