+++ app/api/graphql/schema/queries/query.py
@@ -1,0 +1,9 @@
+import strawberry
+
+from ..types.type import User
+
+@strawberry.type
+class Query:
+    @strawberry.field
+    def user(self) -> User:
+        return User(name="Patrick", age=100)
