+++ app/api/graphql/schema/types/type.py
@@ -1,0 +1,6 @@
+import strawberry
+
+@strawberry.type
+class User:
+    name: str
+    age: int
